Example client: Rename `name` to `queueName` to make the usage clear and explicit
The client example uses a variable called `name`.
Reading through example, questions will rise "What name is it?".
Especially in the context of AMQP/RabbitMQ this can be the Exchange name,
the queue name, the routing key name, the consumer name, etc.

To make the usage clear and explicit, we rename `name` to `queueName`.