tests: add testcase to ensure reader routine terminates
Add a testcase for the bug that the reader go-routine tries to send a
message to the buffered rpc channel but call() terminated because it
read an error from the errors chan or the errors chan was closed.
It cause that reader routine gets stuck forever and does not terminate
when the connection is closed.
More information: https://github.com/rabbitmq/amqp091-go/issues/69.

This testcase does not reproduce the issue reliably, but it is triggered
in ~80% of executions.