ci: run tests via make
Run the tests in github actions via "make tests" instead of duplicating
the command to run the tests.

This makes it easier to run the tests the same way locally and keep the
commands in sync.

The makefile tests target is changed to pass "-cpu=1,2", to run the
tests in CI the same way then before.

ci: remove travis ci config

Travis is not used anymore, github actions is used instead

Skip flaky TestRepeatedChannelExceptionWithPublishAndMaxProcsIssue46 test in CI