ci: run golangci-lint
Add a github action workflow to run golangci-lint.
golangci-lint is configured to only show new issues from PRs.

When all existing issues in the main branch are fixed (or excluded),
only-new-issues could be set to false.

ci: add golangci.yml config file

add a config file to be able to include the integrationtests, this
prevents that code used only in integrationtests is reported as unused