channel/recv: define default case inside switch statement
Multiple channel recv* methods consist of switch-case statement and a
panic() that is called outside the switch when no case matched.

Define instead a default case, this prevents that we have to return
inside each case-statement to prevent that the panic will be called.
This will help to simplify the code in a following commit.