Add convenience function to set connection name
RabbitMQ reads the connection property "connection_name" from the client
and displays the value in the Management UI. This is convenient for
debugging and locating specific client connections, and correlating a
connection to an app.

Showed in the examples how to use this function and added an example in
Go Docs.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>