Add documentation for Connection.UpdateSecret
Setting up an OAuth2 server to end-2-end test this function feels a bit
overkill. This function is used mainly in the context of OAuth2
authentication, where tokens expire. From this client perspective, all
it has to ensure is that a specific frame is put out in the wire to
RabbitMQ server. Then we trust that RabbitMQ will hold its part of the
contract and do the right thing.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>