Fix example client to avoid deadlock in Close (#123)
* Fix example client to avoid deadlock in Close

The example was "giving up" on publish confirmation after the
resendDelay. This is not correct because the library will keep track of
this confirmation, and it will try to deliver the confirmation. By
giving up, we are may leave confirmations un-received in the
confirmation channel, which will cause a deadlock during the shut down
sequence in Channel.Close.

We should not give up on the confirmation and simply wait.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Fix golangci-lint error

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>
Co-authored-by: Luke Bakken <luke@bakken.io>