Add CloseDeadline function to Connection
A deadline to set an I/O timeout to Connection closure is important,
because RabbitMQ may block the connection (during a memory alarm), and
stop reading from the connection socket. In this case, Connection.Close()
will block on I/O. Whilst closing the connection is not the solution to
the memory alarm, an application should have the ability to close a
connection to a service, regardless of the state of the service.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>