Unwind Consume function
After #192, Consume was calling the new ConsumeWithContext function with
a context.Background context. Whilst this would work, it would also pile
up routines that would only return when the AMQP channel is closed. This
is unnecessary, and some code duplication is preferrable over piling up
blocked routines.

Some clarifications to function documentation, and added new paragraphs
to ConsumeWithContext function to explain the semantics.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>