Update deps: use Go 1.20
In #251, we used `URI.Has()` function, which was introduced in Go 1.17.
We also have a dependency on uber-go-leak, which requires Go 1.20. In
addition, we only test in "oldstable" and "stable" versions in CI. That
means, we test in version N and N-1, where N is the latest available
version of Go.

Signed-off-by: Aitor Perez Cedres <aitor.perez@broadcom.com>