Un-deprecate non-context Publish functions
The context was not honoured in any of the *WithContext functions. This
is confusing, and arguably broken. However, we cannot immediately fix
the context-support situation due to https://github.com/rabbitmq/amqp091-go/issues/124#issuecomment-1578529925

This commit undeprecates the non-context variants of publish, and
documents that both variants are equivalent. The example now favours the
non-context variants.

Related to #195

Signed-off-by: Aitor Perez Cedres <aitor.perez@broadcom.com>